More matrix multiplication
Former-commit-id: d444044fd5da2ef79ca3dec5598c798ae6de5d56 [formerly 49588a718e569a25f4635fcc05ba506b46057074]
Former-commit-id: 838241c30cdb2b1935f99ad4657608c80c596d44

+++ src/librapid/config.hpp
@@ -176,12 +176,12 @@ namespace librapid {
 
 	inline void setNumThreads(int64_t num) {
 #if defined(LIBRAPID_HAS_OPENBLAS)
-		openblas_set_num_threads(num);
-		goto_set_num_threads(num);
+		openblas_set_num_threads((int) num);
+		goto_set_num_threads((int) num);
 #endif
 
 #if defined(_OPENMP)
-		omp_set_num_threads(num);
+		omp_set_num_threads((int) num);
 #endif
 	}
